+++ NEWS.md
@@ -256,3 +256,4 @@ Then, we fixed this by using the canonical URL of the CRAN package "https://cran
 ## Major changes (new functions, etc...)
 - We fixed the error in `calcGRM` function with `methodGRM == "linear` option.
 - We fixed the error in `estPhylo` and `plotPhyloTree` functions when drawing phylogenetic trees.
+- We fixed the error in `estNetwork` and `plotHaploNetwork` functions when drawing haplotype network.